Add files via upload
The custom algorithm I made namely NQueenSolver also happens to be very optimal in comparison to the rest. Since this game is supposed to be a study of difference in algorithms and their speeds, it also gives the time taken to generate the relevant solutions for the algorithm so that it may be visually proven how different these algorithms are.